+++ src/main/java/com/calculadora/calculadoraseguro/usecase/CalcularPrecoSeguro.java
@@ -1,16 +1,18 @@
 package com.calculadora.calculadoraseguro.usecase;
 
 import com.calculadora.calculadoraseguro.exception.PrecoBaseException;
-import com.calculadora.calculadoraseguro.adapters.implemetation.CalculoCOFINS;
-import com.calculadora.calculadoraseguro.adapters.implemetation.CalculoIOF;
-import com.calculadora.calculadoraseguro.adapters.implemetation.CalculoPIS;
+import com.calculadora.calculadoraseguro.usecase.adapters.implemetation.CalculoCOFINS;
+import com.calculadora.calculadoraseguro.usecase.adapters.implemetation.CalculoIOF;
+import com.calculadora.calculadoraseguro.usecase.adapters.implemetation.CalculoPIS;
 import com.calculadora.calculadoraseguro.gateway.entity.SeguroCategoria;
 import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
 
 @Service
 public class CalcularPrecoSeguro {
 
-    public double executar(Double precoBase, SeguroCategoria seguro) {
+    public BigDecimal executar(BigDecimal precoBase, SeguroCategoria seguro) {
         CalcularPreco calculadora = new CalcularPreco(
                 new CalculoIOF(seguro.getIof()),
                 new CalculoPIS(seguro.getPis()),
